// lumin-web-backend-2.0/src/CustomRules/custom-rule.enum.ts
export const enum CustomRuleAction {
  USE_S3_STORAGE = 'USE_S3_STORAGE',
  REQUEST_ACCESS_DOCUMENT = 'REQUEST_ACCESS_DOCUMENT',
  CREATE_ORGANIZATION = 'CREATE_ORGANIZATION',
  UPLOAD_THIRD_PARTY_DOCUMENTS = 'UPLOAD_THIRD_PARTY_DOCUMENTS',
  MOVE_DOCUMENTS = 'MOVE_DOCUMENTS',
  ACCESS_DOCUMENT = 'ACCESS_DOCUMENT',
  SHARE_DOCUMENT = 'SHARE_DOCUMENT',
  INVITE_MEMBER_TO_ORGANIZATION = 'INVITE_MEMBER_TO_ORGANIZATION',
  SEND_REQUEST_JOIN_ORG = 'SEND_REQUEST_JOIN_ORG',
  CREATE_ORGANIZATION_BY_ADMIN = 'CREATE_ORGANIZATION_BY_ADMIN',
  VALIDATE_IP_WHITELIST = 'VALIDATE_IP_WHITELIST',
  GET_PROMPT_DRIVE_USERS = 'GET_PROMPT_DRIVE_USERS',
  CHANGE_USER_EMAIL = 'CHANGE_USER_EMAIL',
  ONLY_INTERNAL_INVITE = 'ONLY_INTERNAL_INVITE',
  MANAGE_DOCUMENT_TEMPLATE = 'MANAGE_DOCUMENT_TEMPLATE',
  RESTRICTED_FROM_UPLOADING_DOCUMENT = 'RESTRICTED_FROM_UPLOADING_DOCUMENT',
}